Add support for timezone offsets on a per channel basis.
You can now create a file at CHANNEL_LOCATIONS_FILE that contains a
list of channels and (pytz) timezone location names.
All log files for specified channels from that point on will have their
timestamps adjusted to the given timezone.
If a channel is not specified in the config file, it will default to
the timezone specified by DEFAULT_TIMEZONE (which defaults to UTC).

I've started storing this info in a separate config file as I plan to
move all the other settings into a config file as well so people don't
need to edit the script (allowing the script to be re-used more easily).